put Context in it's own file, remove from Bot. fix imports and tests
