+++ src/lib.rs
@@ -1,11 +1,13 @@
-pub use bot::{Bot, Context, Stepable};
+pub use bot::{Bot, Stepable};
 pub use client_settings::ClientSettings;
+pub use context::Context;
 pub use errors::StepError;
 pub use http_requester::HttpRequester;
 pub use request::Request;
 
 mod bot;
 mod client_settings;
+mod context;
 mod errors;
 mod http_requester;
 mod request;
